requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore==1.6.0
PyJWT
